+++ .jpb/persistence-units.xml
@@ -4,7 +4,7 @@
     <persistence-units>
       <persistence-unit name="Default">
         <packages>
-          <package value="com.example.agenteD" />
+          <package value="com.example.MonitorAgent" />
         </packages>
       </persistence-unit>
     </persistence-units>
