+++ src/main/java/com/example/MonitorAgent/ComponentInit/ComponentInit.java
@@ -1,13 +1,17 @@
-package com.example.agenteD.ComponentInit;
+package com.example.MonitorAgent.ComponentInit;
 
-import com.example.agenteD.Entity.*;
-import com.example.agenteD.Refractory.EntityRefractory;
+import com.example.MonitorAgent.Entity.*;
+import com.example.MonitorAgent.Refractory.EntityRefractory;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.CommandLineRunner;
 import org.springframework.stereotype.Component;
 
+import java.security.PrivateKey;
+import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.List;
 import java.util.concurrent.CompletableFuture;
 
@@ -18,6 +22,8 @@ public class ComponentInit implements CommandLineRunner {
     @Autowired
     EntityRefractory entityRefractory;
 
+    @Value("${monitor.application}")
+    private ArrayList<String> application;
     public void run(String... args) throws Exception {
 
         CompletableFuture<List<Api>> apiResponse = entityRefractory.apiCompletableFuture();
@@ -27,7 +33,9 @@ public class ComponentInit implements CommandLineRunner {
         CompletableFuture<List<Persistence>> persistenceResponse = entityRefractory.persistenceCompletableFuture();
         CompletableFuture<List<Service>> serviceResponse = entityRefractory.serviceCompletableFuture();
 
-
+logger.info("{}",application);
+int num = application.size();
+logger.info("{}",num);
             CompletableFuture.allOf(apiResponse, applicationResponse, integrationResponse,
                     loadBalancerResponse, persistenceResponse, serviceResponse).join();
