+++ src/main/java/com/example/MonitorAgent/Entity/Service.java
@@ -1,4 +1,4 @@
-package com.example.agenteD.Entity;
+package com.example.MonitorAgent.Entity;
 
 import lombok.*;
 
@@ -10,31 +10,34 @@ import java.time.LocalDate;
 @NoArgsConstructor
 @Getter
 @Setter
-@Entity(name = "Api")
-@Table(name = "api")
-public class Api {
+@Entity(name = "Service")
+@Table(name = "service")
+public class Service {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "api_id", nullable = false)
-    private Integer api_id;
-
-    @Column(name = "status", length = 45)
-    private String status;
-
-    @Column(name = "application_id")
-    private Integer applicationId;
-
-    @Column(name = "common")
-    private Boolean common;
+    @Column(name = "service_id", nullable = false)
+    private Integer service_id;
 
     @Column(name = "description", length = 45)
     private String description;
 
+    @Column(name = "application_id")
+    private Integer applicationId;
+
     @Column(name = "service_name", length = 45)
     private String serviceName;
 
-    @Column(name = "name_space", length = 45)
-    private String nameSpace;
+    @Column(name = "operation", length = 45)
+    private String operation;
+
+    @Column(name = "port", length = 45)
+    private String port;
+
+    @Column(name = "method", length = 45)
+    private String method;
+
+    @Column(name = "test_url", length = 45)
+    private String testUrl;
 
     @Column(name = "num_test")
     private Integer numTest;
@@ -45,11 +48,11 @@ public class Api {
     @Column(name = "hist_failed_test")
     private Long histFailedTest;
 
-    @Column(name = "min_test_failed")
-    private Integer minTestFailed;
+    @Column(name = "max_failed_test")
+    private Integer maxFailedTest;
 
-    @Column(name = "max_test_failed")
-    private Integer maxTestFailed;
+    @Column(name = "min_failed_test")
+    private Integer minFailedTest;
 
     @Column(name = "low_alarm")
     private Integer lowAlarm;
@@ -57,19 +60,19 @@ public class Api {
     @Column(name = "high_alarm")
     private Integer highAlarm;
 
-    @Column(name = "last_test_date")
-    private LocalDate lastTestDate;
-
     @Column(name = "\"req/seg\"", length = 45)
     private String reqSeg;
 
     @Column(name = "test_interv")
     private Long testInterv;
 
-    @Column(name = "consecutive_successful_test")
-    private Integer consecutiveSuccessfulTest;
-
     @Column(name = "hist_successful_test")
     private Long histSuccessfulTest;
+
+    @Column(name = "last_tests_date")
+    private LocalDate lastTestsDate;
+
+    @Column(name = "status", length = 45)
+    private String status;
 
 }
