+++ src/main/java/com/example/MonitorAgent/MonitorAgent.java
@@ -1,13 +1,13 @@
-package com.example.agenteD;
+package com.example.MonitorAgent;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class AgentMonitorApplication {
+public class MonitorAgent {
 
 	public static void main(String[] args) throws InterruptedException {
-		SpringApplication.run(AgentMonitorApplication.class, args);
+		SpringApplication.run(MonitorAgent.class, args);
 	}
 
 }
