+++ src/main/java/com/example/MonitorAgent/Repository/ApiRepository.java
@@ -1,11 +1,8 @@
-package com.example.agenteD.Repository;
+package com.example.MonitorAgent.Repository;
 
-import com.example.agenteD.Entity.Api;
+import com.example.MonitorAgent.Entity.Api;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-
-import java.util.List;
-import java.util.Set;
 
 @Repository
 public interface ApiRepository extends JpaRepository<Api, Integer> {
