+++ src/main/java/com/example/MonitorAgent/Repository/ApplicationRepository.java
@@ -1,6 +1,6 @@
-package com.example.agenteD.Repository;
+package com.example.MonitorAgent.Repository;
 
-import com.example.agenteD.Entity.Application;
+import com.example.MonitorAgent.Entity.Application;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
