+++ src/main/java/com/example/MonitorAgent/Repository/ServiceRepository.java
@@ -1,6 +1,6 @@
-package com.example.agenteD.Repository;
+package com.example.MonitorAgent.Repository;
 
-import com.example.agenteD.Entity.Service;
+import com.example.MonitorAgent.Entity.Service;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
