+++ src/main/resources/application.properties
@@ -13,3 +13,5 @@ spring.jpa.properties.hibernate.dialect= org.hibernate.dialect.PostgreSQLDialect
 
 spring.jpa.generate-ddl= false
 #spring.jpa.hibernate.ddl-auto = update
+
+monitor.application = BDVenLineaPersonas,BDVdigital
